Remove unit tests
Unfortuntely there seems to be some Fabric2Forge bug that is
preventing one (or both I can't recall) of those classes from being
remapped.

Fabric2Forge does include a stage for the test source set, and I would
like to have these in future - so this is only intended as a temporary
patch.