Adapt mixins for Minecraft Forge
This will be merged with my commit on the other branch, I forgot
to cherry-pick that to this branch.

This is also currently incomplete, further work needs to be done -
it seems the latest version collided with Minecraft Forge than I
needed to fix for the last... or maybe my memory is waining, or I'm
just doing a better job at porting this time around.

I'll probably return to this tommorrow to finish it up and make a
release :)